==> scripts/05b_FINAL_models.Rmd <==
---
title: "05b_FINAL_models_and_plots"
output: html_document
date: "2024-06-04"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Packages
```{r}

# reinstall matrix from source (if needed)
#installed.packages() |> as.data.frame() |> subset(Package=="Matrix", select = c(LibPath, Version))
remotes::install_version("Matrix", version = "1.6-3")

#library(pacman)
#install.packages("TMB",type="source")
installed.packages() |> as.data.frame() |> subset(Package=="Matrix", select=c(Package, LibPath, Version))

#install.packages("glmmTMB")

install.packages("Matrix", type = "source")
install.packages("TMB", type = "source")
install.packages("RccpEigen", type = "source")
install.packages("glmmTMB", type = "source")
library("glmmTMB")

library(pacman)
p_load(corrplot,glmmTMB,AICcmodavg,sjPlot,DHARMa,MuMIn,ggplot2,janitor,dplyr,tidyr,
       readr,here,viridis)



```

## Read in data 
```{r}

# our new make names function
make_names <- function(x) {
  gsub("\\.", "_", make.names(gsub("*.csv$", "", x)))
}


# read in count and covariate data -- both cities
all_dat <- read_csv(here("data", "covariates", "WIDE_COUNTS_ALL_ENV_URB_SITES_1000m.csv")) %>%
  subset(season != "Spring20" & 
           season != "Fall20" & 
           season != "Summer20") %>%# %>% keep only win19-win20
  # subset(site != "P-OV1" & site != "P-TE1" & site != "P-RWSR" & site != "P-OLT")
#   clean_names() # make spaces underscores %>% 
 rename_with(janitor::make_clean_names, all_of(1:23)) %>% 
  filter(days_active != 0)# make spaces underscores for spp names (without janitor) %>%


all_dat_sewa <- read_csv(here("data", "covariates", "SEWA_WIDE_COUNTS_ALL_ENV_URB_SITES_1000m.csv")) %>%
  subset(season != "Spring20" & 
           season != "Fall20" & 
           season != "Summer20") %>%# %>% keep only win19-win20
  # subset(site != "P-OV1" & site != "P-TE1" & site != "P-RWSR" & site != "P-OLT")
#   clean_names() # make spaces underscores %>% 
 rename_with(janitor::make_clean_names, all_of(1:23)) %>% # make spaces underscores for spp names (without janitor) 
filter(days_active != 0)

all_dat_tawa <- read_csv(here("data", "covariates", "TAWA_WIDE_COUNTS_ALL_ENV_URB_SITES_1000m.csv")) %>%
  subset(season != "Spring20" & 
           season != "Fall20" & 
           season != "Summer20") %>%# %>% keep only win19-win20
  # subset(site != "P-OV1" & site != "P-TE1" & site != "P-RWSR" & site != "P-OLT")
#   clean_names() # make spaces underscores %>% 
 rename_with(janitor::make_clean_names, all_of(1:23)) %>% # make spaces underscores for spp names (without janitor)
  filter(days_active != 0)

# read in richness/diversity data 
all_dat_veg <- read_csv(here("data", "covariates", "vegan_sites_all_covs.csv")) 
all_dat_veg_sewa <- read_csv(here("data", "covariates", "sewa_vegan_sites_all_covs.csv")) 
all_dat_veg_tawa <- read_csv(here("data", "covariates", "tawa_vegan_sites_all_covs.csv")) 

# make things factors
all_dat$city <- as.factor(all_dat$city)
all_dat$season <- as.factor(all_dat$season)
all_dat$site <- as.factor(all_dat$site)

all_dat_sewa$city <- as.factor(all_dat_sewa$city)
all_dat_sewa$season <- as.factor(all_dat_sewa$season)
all_dat_sewa$site <- as.factor(all_dat_sewa$site)

all_dat_tawa$city <- as.factor(all_dat_tawa$city)
all_dat_tawa$season <- as.factor(all_dat_tawa$season)
all_dat_tawa$site <- as.factor(all_dat_tawa$site)

# make things factors
all_dat_veg_sewa$city <- as.factor(all_dat_veg_sewa$city)
all_dat_veg_tawa$site <- as.factor(all_dat_veg_tawa$site)

# data will be scaled in model so we don't need to do it now
```

## Variable distribution plots 
```{r}
# let's create a data set with only 1 row per site
all_cov <- all_dat %>% distinct(site, .keep_all = TRUE) %>%
    mutate(city = as.character(city)) %>% 
    mutate(city = dplyr::recode(city, sewa = "Seattle",tawa = "Tacoma")) # rename cities for better labels

# look at pollution burden distributions
burden_plot <- all_cov %>%
  ggplot(aes(x = BURDEN_PERCENTILE)) +
  geom_histogram(fill = "white", colour = "black") +
  facet_grid(city ~ .)

impervious_veg <- all_cov %>%
  ggplot(aes(x = Impervious, y = prop_ndvi_avg)) +
  geom_point(color = "red") +  # Set point color
  labs(x = "impervious", y = "NDVI") 


# let's create a box plot showing the distribution of our variables
# for each city 

# environmental exposures  


```

## Correlation plots 
```{r}

# let's rename all variables for ease of plotting correlations


## Correlation plots 
colnames(all_dat)
cordat <- all_dat %>% distinct(site, .keep_all = TRUE) %>% 
  mutate(city = as.character(city)) %>% 
  mutate(city = dplyr::recode(city, sewa = "Seattle",tawa = "Tacoma")) %>%
  dplyr::select(c(
    natural,
     # urban,
     Impervious,
     # prop_ndvi_avg,
     POPDEN2020,
     med_income,
     prox_heavy_traffic,
     pm2.5_count,
     diesel_tons_km2,
     avg_rsei,
     avg_ozone,
     avg_ptsdf,
     units_lead,
     avg_pnpl,
     avg_prmp
    #     "Wastewater discharge" = avg_pwdis
  )) %>%
  rename("Natural land cover" = natural ,
     #    "Urban land cover (NLCD)" = urban,
         "Median income" = med_income,
         "Impervious surface" = Impervious,
     #    "Average NDVI" = prop_ndvi_avg,
         "Human pop. density" = POPDEN2020,
         "Prox. to heavy traffic roadways*" = prox_heavy_traffic,
         "PM2.5 concentration*" = pm2.5_count,
         "Diesel/NOx emissions (tons/km2)*" = diesel_tons_km2,
         "Toxic releases*" = avg_rsei,
         "Average ozone concentration*" = avg_ozone,
         "Prox. to hazardous waste facilities**" = avg_ptsdf,
         "Units with lead**" = units_lead,
         "Prox. to Superfund sites**" = avg_pnpl,
         "Prox. to risk mgmt plan facilities**" = avg_prmp
    #     "Wastewater discharge" = avg_pwdis
  ) 
  

print(colnames(cordat))

corr <- cor(cordat, method = "spearman", use="pairwise.complete.obs")

corrplot(corr, 
         method = 'color', 
         type = 'lower', 
         order = 'original', 
         tl.col = 'black',
         cl.ratio = 0.2, 
         tl.srt = 45, 
         col = viridis(200),  # Use viridis for colorblind-friendly palette
         cl.pos = 'b', 
         addgrid.col = 'white', 
         addCoef.col = 'black')  # Coefficient text color

##### Variables grouped 

cordat <- all_dat %>% distinct(site, .keep_all = TRUE) %>% 
  mutate(city = as.character(city)) %>% 
  mutate(city = dplyr::recode(city, sewa = "Seattle",tawa = "Tacoma")) %>%
  dplyr::select(c(natural,
                  med_income,
                  Impervious, 
                  POPDEN2020,
                  EXPOSURE,
                  RISK,
                  BURDEN_PERCENTILE)) %>%
  rename("Natural land cover" = natural,
  "Median income" = med_income,
  "Impervious surface" = Impervious,
  "Human pop. density" = POPDEN2020,
  "Environmental exposures" = EXPOSURE,
  "Environmental risks" =  RISK,
  "Pollution burden" = BURDEN_PERCENTILE)

corr <- cor(cordat, method = "spearman", use="pairwise.complete.obs")

corrplot(corr, 
         method = 'color', 
         type = 'lower', 
         order = 'original', 
         tl.col = 'black',
         cl.ratio = 0.2, 
         tl.srt = 45, 
         col = viridis(200),  # Use viridis for colorblind-friendly palette
         cl.pos = 'b', 
         addgrid.col = 'white', 
         addCoef.col = 'black')  # Coefficient text color

#### look at correlations -- by city

##  seattle -- all

cordat <- all_dat_sewa %>% distinct(site, .keep_all = TRUE) %>% 
  dplyr::select(c(
    natural,
     # urban,
     Impervious,
     # prop_ndvi_avg,
     POPDEN2020,
     med_income,
     prox_heavy_traffic,
     pm2.5_count,
     diesel_tons_km2,
     avg_rsei,
     avg_ozone,
     avg_ptsdf,
     units_lead,
     avg_pnpl,
     avg_prmp
    #     "Wastewater discharge" = avg_pwdis
  )) %>%
  rename("Natural land cover" = natural ,
     #    "Urban land cover (NLCD)" = urban,
         "Median income" = med_income,
         "Impervious surface" = Impervious,
     #    "Average NDVI" = prop_ndvi_avg,
         "Human pop. density" = POPDEN2020,
         "Prox. to heavy traffic roadways*" = prox_heavy_traffic,
         "PM2.5 concentration*" = pm2.5_count,
         "Diesel/NOx emissions (tons/km2)*" = diesel_tons_km2,
         "Toxic releases*" = avg_rsei,
         "Average ozone concentration*" = avg_ozone,
         "Prox. to hazardous waste facilities**" = avg_ptsdf,
         "Units with lead**" = units_lead,
         "Prox. to Superfund sites**" = avg_pnpl,
         "Prox. to risk mgmt plan facilities**" = avg_prmp
    #     "Wastewater discharge" = avg_pwdis
  ) 

corr <- cor(cordat, method = "spearman", use="pairwise.complete.obs")
corrplot(corr, method = "number")

corrplot(corr, 
         method = 'color', 
         type = 'lower', 
         order = 'original', 
         tl.col = 'black',
         cl.ratio = 0.2, 
         tl.srt = 45, 
         col = viridis(200),  # Use viridis for colorblind-friendly palette
         cl.pos = 'b', 
         addgrid.col = 'white', 
         addCoef.col = 'black')

## seattle -- grouped 
cordat <- all_dat_sewa %>% distinct(site, .keep_all = TRUE) %>% 
  dplyr::select(c(natural,
                  med_income,
    #              Impervious, 
                  POPDEN2020,
                  SEWA_EXPOSURE,
                  SEWA_RISK,
                  SEWA_BURDEN_PERCENTILE)) %>%
  rename("Natural land cover" = natural,
  "Median income" = med_income,
#  "Impervious surface" = Impervious,
  "Human pop. density" = POPDEN2020,
#  "Environmental exposures" = EXPOSURE,
 # "Environmental risks" =  RISK,
  "Pollution burden" = SEWA_BURDEN_PERCENTILE)

corr <- cor(cordat, method = "spearman", use="pairwise.complete.obs")

corrplot(corr, 
         method = 'color', 
         type = 'lower', 
         order = 'original', 
         tl.col = 'black',
         cl.ratio = 0.2, 
         tl.srt = 45, 
         col = viridis(200),  # Use viridis for colorblind-friendly palette
         cl.pos = 'b', 
         addgrid.col = 'white', 
         addCoef.col = 'black')  # Coefficient text color

## tacoma -- all 
cordat <- all_dat_tawa %>% distinct(site, .keep_all = TRUE) %>% 
  dplyr::select(c(
    natural,
     # urban,
    # Impervious,
     # prop_ndvi_avg,
     POPDEN2020,
     med_income,
     prox_heavy_traffic,
     pm2.5_count,
     diesel_tons_km2,
     avg_rsei,
     avg_ozone,
     avg_ptsdf,
     units_lead,
     avg_pnpl,
     avg_prmp
    #     "Wastewater discharge" = avg_pwdis
  )) %>%
  rename("Natural land cover" = natural ,
     #    "Urban land cover (NLCD)" = urban,
         "Median income" = med_income,
       #  "Impervious surface" = Impervious,
     #    "Average NDVI" = prop_ndvi_avg,
         "Human pop. density" = POPDEN2020,
         "Prox. to heavy traffic roadways*" = prox_heavy_traffic,
         "PM2.5 concentration*" = pm2.5_count,
         "Diesel/NOx emissions (tons/km2)*" = diesel_tons_km2,
         "Toxic releases*" = avg_rsei,
         "Average ozone concentration*" = avg_ozone,
         "Prox. to hazardous waste facilities**" = avg_ptsdf,
         "Units with lead**" = units_lead,
         "Prox. to Superfund sites**" = avg_pnpl,
         "Prox. to risk mgmt plan facilities**" = avg_prmp
    #     "Wastewater discharge" = avg_pwdis
  ) 

corr <- cor(cordat, method = "spearman", use="pairwise.complete.obs")

corrplot(corr, 
         method = 'color', 
         type = 'lower', 
         order = 'original', 
         tl.col = 'black',
         cl.ratio = 0.2, 
         tl.srt = 45, 
         col = viridis(200),  # Use viridis for colorblind-friendly palette
         cl.pos = 'b', 
         addgrid.col = 'white', 
         addCoef.col = 'black')



## tacoma -- grouped 
cordat <- all_dat_tawa %>% distinct(site, .keep_all = TRUE) %>% 
  dplyr::select(c(natural,
                  med_income,
    #              Impervious, 
                  POPDEN2020,
                  TAWA_EXPOSURE,
                  TAWA_RISK,
                  TAWA_BURDEN_PERCENTILE)) %>%
  rename("Natural land cover" = natural,
  "Median income" = med_income,
#  "Impervious surface" = Impervious,
  "Human pop. density" = POPDEN2020,
#  "Environmental exposures" = EXPOSURE,
 # "Environmental risks" =  RISK,
  "Pollution burden" = TAWA_BURDEN_PERCENTILE)

corr <- cor(cordat, method = "spearman", use="pairwise.complete.obs")

corrplot(corr, 
         method = 'color', 
         type = 'lower', 
         order = 'original', 
         tl.col = 'black',
         cl.ratio = 0.2, 
         tl.srt = 45, 
         col = viridis(200),  # Use viridis for colorblind-friendly palette
         cl.pos = 'b', 
         addgrid.col = 'white', 
         addCoef.col = 'black')  # Coefficient text color

```

## Black bear models 
```{r}

# Grouped exposures/effects model
bearmod <- glmmTMB(black_bear ~ city + 
                     scale(natural) + 
                   #  scale(EXPOSURE) + 
                     scale(RISK)  + 
                     scale(med_income) + season + 
                     scale(POPDEN2020) + ( 1 | site), 
                   data = all_dat, 
                   offset = log(days_active),
                   family = "nbinom2",
                   na.action = "na.fail")
summary(bearmod)

bearmodsewa <- glmmTMB(Black_bear ~ scale(develop) + 
                     scale(EXPOSURE) + scale(RISK)  + 
                     scale(med_income) + season + 
                     # scale(pop_density) 
                     + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days.active),
                family = "nbinom2")

bearmodtawa <- glmmTMB(Black_bear ~ scale(develop) + 
                     scale(EXPOSURE) + scale(RISK)  + 
                     scale(med_income) + season + scale(pop_density) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days.active),
                family = "nbinom2")

# None of the models converge

```


## Bobcat models 
```{r}

# Grouped exposures/effects model
bobcatmod <- glmmTMB(bobcat ~ city + 
                       scale(natural) + 
           #          scale(EXPOSURE) + 
                       scale(RISK)  + 
                     scale(med_income) + season + 
             scale(POPDEN2020) + ( 1 | site), 
                   data = all_dat, 
                   offset = log(days_active),
                   family = "nbinom2",
                   na.action = "na.fail")
summary(bobcatmod)

bobcatmodsewa <- glmmTMB(bobcat ~ scale(natural) + 
     #                scale(SEWA_EXPOSURE) + 
        #               scale(SEWA_RISK)  + 
                     scale(med_income) + season + 
                       scale(POPDEN2020) 
                       + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days_active),
                family = "nbinom2")

bobcatmodtawa <- glmmTMB(bobcat ~ scale(natural) + 
                     scale(TAWA_EXPOSURE) + scale(TAWA_RISK)  + 
                     scale(med_income) + season + 
                       scale(POPDEN2020) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days_active),
                family = "nbinom2")

# model diagnostics 
simOut <- simulateResiduals(fittedModel = bobcatmod, plot = T)

testOutliers(bobcatmod)

testDispersion(bobcatmod)

testZeroInflation(coymodglobal) # test for zero-inflation -- In a model with variable dispersion, 
# this is often the signal that some other distributional assumptions are violated, 
# but not seeing zero-inflation so we can probably  move on with this model 

plotResiduals(simOut, form = all_dat$site)
plot(simOut, quantreg = T)
par(mfrow = c(1,2))
plotResiduals(simOut, all_dat$city) 

plotResiduals(simOut, all_dat$Environment1, quantreg = T)


# Model converges for global model but not city specific models 
```


## Coyote models
```{r}
# Coyote model -- negative binomial, grouped exposures and effects
colnames(all_dat)
coymod <- glmmTMB(coyote ~ city +
                    scale(natural) + 
                    # scale(EXPOSURE) + 
                    scale(RISK)  + 
                     scale(med_income) + season + scale(POPDEN2020) + ( 1 | site), 
                data = all_dat, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")

summary(coymod)

coymodsewa <- glmmTMB(coyote ~ 
                        scale(natural) + 
                #     scale(EXPOSURE) + 
                       scale(SEWA_RISK)  + 
                     scale(med_income) + 
                  season + scale(POPDEN2020) + 
                       + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days_active),
                family = "nbinom2",na.action = "na.fail")

summary(coymodsewa)

coymodtawa <- glmmTMB(coyote ~ 
                        scale(natural) + 
                   #  scale(EXPOSURE) + 
                       scale(TAWA_RISK)  + 
                     scale(med_income) + season + scale(POPDEN2020) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")

summary(coymodtawa)
```


## Raccoon models
```{r}

# Grouped exposures/effects model
racmod <- glmmTMB(raccoon ~ city + 
                    scale(natural) + 
                   #  scale(EXPOSURE) + 
                    scale(RISK)  + 
                     scale(med_income) + season + scale(POPDEN2020) + ( 1 | site), 
                   data = all_dat, 
                   offset = log(days_active),
                   family = "nbinom2",
                   na.action = "na.fail")
summary(racmod)

racmodsewa <- glmmTMB(raccoon ~ scale(natural) + 
                    # scale(SEWA_EXPOSURE) + 
                       scale(SEWA_RISK)  + 
                     scale(med_income) + season +
                        scale(POPDEN2020)
                       + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")

summary(racmodsewa)

racmodtawa <- glmmTMB(raccoon ~ scale(natural) + 
                  #   scale(TAWA_EXPOSURE) + 
                       scale(TAWA_RISK)  + 
                     scale(med_income) + season + 
                       scale(POPDEN2020) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")
summary(racmodtawa)

```


## River otter models 
```{r}

# Grouped exposures/effects model
ottermod <- glmmTMB(River_otter ~ city + scale(develop) + 
                     scale(EXPOSURE) + scale(RISK)  + 
                     scale(med_income) + season + scale(pop_density) + ( 1 | site), 
                   data = all_dat, 
                   offset = log(days.active),
                   family = "nbinom2",
                   na.action = "na.fail")
summary(ottermod)

ottermodsewa <- glmmTMB(River_otter ~ scale(develop) + 
                     scale(EXPOSURE) + scale(RISK)  + 
                     scale(med_income) + season + 
                       # scale(pop_density) 
                       + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days.active),
                family = "nbinom2")

ottermodtawa <- glmmTMB(River_otter ~ scale(develop) + 
                     scale(EXPOSURE) + scale(RISK)  + 
                     scale(med_income) + season + scale(pop_density) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days.active),
                family = "nbinom2")

# None of the models converge 

```

## Striped skunk models 
```{r}

# Grouped exposures/effects model
skunkmod <- glmmTMB(striped_skunk ~ city + 
                      scale(natural) + 
                   #  scale(EXPOSURE) + 
                      scale(RISK)  + 
                     scale(med_income) + season + 
                     scale(POPDEN2020) + ( 1 | site), 
                   data = all_dat, 
                   offset = log(days_active),
                   family = "nbinom2",
                   na.action = "na.fail")
summary(skunkmod)

skunkmodsewa <- glmmTMB(striped_skunk ~ scale(natural) + 
                 #    scale(SEWA_EXPOSURE) + 
                       scale(SEWA_RISK)  + 
                     scale(med_income) + season +
                        scale(POPDEN2020)
                       + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")

summary(skunkmodsewa)

skunkmodtawa <- glmmTMB(striped_skunk ~ 
                          scale(natural) + 
               #      scale(TAWA_EXPOSURE) + 
                       scale(TAWA_RISK)  + 
                     scale(med_income) + season + 
                       scale(POPDEN2020) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")

summary(skunkmodtawa)

# model diagnostics 
simOut <- simulateResiduals(fittedModel = skunkmod, plot = T)

testOutliers(bobcatmod)

testDispersion(skunkmod)

testZeroInflation(skunkmodtawa) # test for zero-inflation -- In a model with variable dispersion, 
# this is often the signal that some other distributional assumptions are violated, 
# but not seeing zero-inflation so we can probably  move on with this model 

plotResiduals(simOut, form = all_dat$site)
plot(simOut, quantreg = T)
par(mfrow = c(1,2))
plotResiduals(simOut, all_dat$city) 

plotResiduals(simOut, all_dat$Environment1, quantreg = T)

```

## Virginia opossum models
```{r}
oppmod <- glmmTMB(virginia_opossum ~ city + 
                    scale(natural) + 
                    # scale(EXPOSURE) + 
                    scale(RISK)  + 
                     scale(med_income) + season + 
                    scale(POPDEN2020) + ( 1 | site), 
                   data = all_dat, 
                   offset = log(days_active),
                   family = "nbinom2",
                   na.action = "na.fail")
summary(oppmod)

oppmodsewa <- glmmTMB(virginia_opossum ~ 
                     scale(natural) + 
             #        scale(SEWA_EXPOSURE) + 
                       scale(SEWA_RISK)  + 
                     scale(med_income) + season +
                        scale(POPDEN2020)
                       + ( 1 | site), 
                data = all_dat_sewa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")
summary(oppmodsewa)

oppmodtawa <- glmmTMB(virginia_opossum ~ 
                        scale(natural) + 
                  #   scale(TAWA_EXPOSURE) + 
                       scale(TAWA_RISK)  + 
                     scale(med_income) + season + 
                       scale(POPDEN2020) + ( 1 | site), 
                data = all_dat_tawa, 
                offset = log(days_active),
                family = "nbinom2",
                na.action = "na.fail")
summary(oppmodtawa)


```


Simpson's diversity model 
```{r}

# Grouped model 
shanmod <- glmmTMB(shannon.di ~ city + 
                     scale(natural) + 
                     #scale(EXPOSURE) + 
                     scale(RISK)  + 
                     scale(med_income) +
                      scale(POPDEN2020) + ( 1 | site), 
                   data = all_dat_veg, 
                   family = "gaussian",
                   na.action = "na.fail") 
summary(shanmod)

shanmodsewa <- glmmTMB(shannon.di ~ 
                         scale(natural) + 
                      #   scale(EXPOSURE) + 
                         scale(SEWA_RISK) + 
                        scale(POPDEN2020) + 
                         scale(med_income) + ( 1 | site), 
                    data = all_dat_veg_sewa, 
                    family = "gaussian")

summary(shanmodsewa)

shanmodtawa <- glmmTMB(shannon.di ~ 
                         scale(natural) + 
                         #scale(EXPOSURE) + 
                         scale(TAWA_RISK) + 
                         scale(POPDEN2020) + 
                         scale(med_income) + ( 1 | site),
                    data = all_dat_veg_tawa, 
                    family = "gaussian")
summary(shanmodtawa)

```


Species richness model 
```{r}
colnames(all_dat_veg)
# Grouped model 
richmod <- glmmTMB(richness ~ city + 
                     scale(natural) + 
                     #scale(EXPOSURE) + 
                     scale(RISK)  + 
                     scale(med_income) +
                      scale(POPDEN2020) + 
                     ( 1 | site), 
                   data = all_dat_veg, 
                   family = "poisson",
                   na.action = "na.fail") 
summary(richmod)

richmodsewa <- glmmTMB(richness ~ 
                         scale(natural) + 
                         #scale(EXPOSURE) + 
                         scale(SEWA_RISK) + scale(POPDEN2020) + 
                         scale(med_income) + ( 1 | site), 
                    data = all_dat_veg_sewa, 
                    family = "poisson")
summary(richmodsewa)

richmodtawa <- glmmTMB(richness ~ 
                         scale(natural) + 
                         #scale(EXPOSURE) + 
                         scale(TAWA_RISK) + 
                         scale(POPDEN2020) + 
                         scale(med_income) + ( 1 | site),
                    data = all_dat_veg_sewa, 
                    family = "poisson")
summary(richmodtawa)

```


## Figures! 

# Results tables 
```{r}

# species models -- both cities 
tab_model(coymod, racmod, oppmod, skunkmod, bobcatmod,
          collapse.ci = TRUE, 
          p.style     = "numeric_stars"# ,
       #    rm.terms = c("season [Fall20,
   #                          Spring19, Spring20, Summer19, Summer20,
    #                        Winter19,Winter20]")
          ) 


# species models -- seattle 
tab_model(coymodsewa, racmodsewa, oppmodsewa, 
          collapse.ci = TRUE, 
          p.style     = "numeric_stars",
          rm.terms = c("season [Fall20,
                            Spring19, Spring20, Summer19, Summer20,
                            Winter19,Winter20]")
          ) 

# species models -- tacoma 
tab_model(coymodtawa, racmodtawa, oppmodtawa, skunkmodtawa,
          collapse.ci = TRUE, 
          p.style     = "numeric_stars",
          rm.terms = c("season [Fall20,
                            Spring19, Spring20, Summer19, Summer20,
                            Winter19,Winter20]")
          ) 


# diversity models 
tab_model(shanmod, shanmodsewa, shanmodtawa,
          collapse.ci = TRUE, 
          p.style     = "numeric_stars",
          rm.terms = c("season [Fall20,
                            Spring19, Spring20, Summer19, Summer20,
                            Winter19,Winter20]")
          ) 

# richness models 
tab_model(richmod, richmodsewa, richmodtawa,
          collapse.ci = TRUE, 
          p.style     = "numeric_stars",
          rm.terms = c("season [Fall20,
                            Spring19, Spring20, Summer19, Summer20,
                            Winter19,Winter20]")
          ) 

```

# Forest plots 
First, both-cities model
```{r}

# Forest plot of species model estimates --grouped models
(plot_global <- plot_models(coymod, racmod, oppmod,
            show.values = T,
            show.legend = F,
            grid = TRUE, vline.color = "gray",
             grid.breaks = c(.1,1,10),
          p.shape     = "FALSE",
         rm.terms = c("season [
                            Winter19,Spring19, Summer19, 
                             Fall19,Winter20]",
                             "city [tawa]") , 
     #       transform = NULL
               axis.lim = c(0.1,10) 
     )+ 
#    scale_y_log10(limits = c(-0.1, 100)) + 
  scale_color_sjplot("eight") + 
  scale_x_discrete(labels=c("Natural land cover", "Environmental risks", "Median income", "Human pop. density"), limits = rev) + 
font_size(axis_title.x = 0, axis_title.y = 12, labels.x = 0, labels.y = 12))


# vline
#state1plot$layers[[1]]$aes_params$size <- 4
# labels
#plot_global$layers[[4]]$aes_params$size <- 3

```


Seattle model
```{r}

# Forest plot of species model estimates -- Seattle 
plot_sewa <- plot_models(coymodsewa, racmodsewa, oppmodsewa,
                         show.values = T, 
                         show.legend = F,
                         grid = TRUE, 
                         axis.labels = NULL,
                         axis.title = NULL,
                         vline.color = "gray", 
                         grid.breaks = c(0.1,1,10),
                         auto.label = FALSE,
                         p.shape = "FALSE",
         rm.terms = c("season [
                            Winter19,Spring19, Summer19, 
                             Fall19,Winter20]") , 
     #       transform = NULL
               axis.lim = c(0,10) 
     )+ 
#    scale_y_log10(limits = c(-0.1, 100)) + 
  scale_color_sjplot("eight") + 
  scale_x_discrete(labels=c("Natural land cover", "Environmental risks", "Median income", "Human pop. density"), limits = rev)+ 
font_size(axis_title.x = 0, axis_title.y = 12, labels.x = 0, labels.y = 12)


```


Tacoma model
```{r}

# Forest plot of species model estimates -- Tacoma
plot_tawa <- plot_models(coymodtawa, racmodtawa, oppmodtawa, 
                         show.values = T, 
                         show.legend = F,
                         grid = TRUE, 
                         axis.labels = NULL,
                         axis.title = NULL,
                         vline.color = "gray", 
                         grid.breaks = c(0.1,1,10),
                         auto.label = FALSE,
          p.shape     = "FALSE",
         rm.terms = c("season [
                            Winter19,Spring19, Summer19, 
                             Fall19,Winter20]") , 
     #       transform = NULL
               axis.lim = c(0,10) 
     )+ 
#    scale_y_log10(limits = c(-0.1, 100)) + 
  scale_color_sjplot("eight") + 
  scale_x_discrete(labels=c("Natural land cover", "Environmental risks", "Median income", "Human pop. density"), limits = rev)+ 
font_size(axis_title.x = 12, axis_title.y = 12, labels.x = 10, labels.y = 12)

```

## Plot all meso models together 
```{r}
cowplot::plot_grid(plot_global,plot_sewa, plot_tawa,
                   labels = c('A', 'B', 'C'), nrow = 3,
          align = "h")
```

Diversity models
```{r}

# Forest plot of species model estimates -- Tacoma
plot_models(simpmod, simpmodsewa, simpmodtawa, title = "Tacoma", show.values = T, grid = TRUE, vline.color = "gray",
            grid.breaks = c(-0.5,0,0.5),
         p.shape     = "FALSE",
         axis.lim = c(-0.5,0.5)
     )+ 
#    scale_y_log10(limits = c(-0.1, 100)) + 
  scale_color_sjplot("eight") + 
  scale_x_discrete(labels=c(("Developed land cover", "Environmental exposures", "Environmental risks", "Median income", "Human pop. density"), limits = rev) 


theme_set(theme_sjplot())
?plot_model
# income plots 
income_both <- plot_model(simpmod, type = "pred", terms = c("med_income"), title = "",
                          axis.title = c("Median income", "Simpson's diversity index"),
           axis.labels = c("$50,000", "$100,000", "$150,000", "$200,000"),
           axis.lim = list(c(48000,175000), c(0.2,0.8)))

income_sewa <- plot_model(simpmodsewa, type = "pred", terms = c("med_income"), title = "",
                          axis.title = c("Median income", ""),
           axis.labels = c("$50,000", "$100,000", "$150,000", "$200,000"),
           axis.lim = list(c(40000,170000), c(0.2,0.8)))

income_tawa <- plot_model(simpmodtawa, type = "pred", terms = c("med_income"), title = "",
                          axis.title = c("Median income", ""),  
                          grid.breaks = c(50000,75000,100000),
           axis.lim = list(c(40000,120000), c(0.2,0.8)))
    

summary(simpmod)
exp_both <- plot_model(simpmod, type = "pred", terms = c("EXPOSURE"), title = "",
                          axis.title = c("Environmental exposure percentile", "Simpson's diversity index"),
                       axis.lim = list(c(0,60), c(0.2,1)))

exp_sewa <- plot_model(simpmodsewa, type = "pred", terms = c("EXPOSURE"), title = "",
                          axis.title = c("Environmental exposure percentile", ""),
                       axis.lim = list(c(0,60), c(0.2,1)))

exp_tawa <- plot_model(simpmodtawa, type = "pred", terms = c("EXPOSURE"), title = "",
                          axis.title = c("Environmental exposure percentile", ""),
                       axis.lim = list(c(0,60), c(0.2,1)))
    

plot_grid(income_both, income_sewa, income_tawa, 
          exp_both, exp_sewa, exp_tawa, labels = c('A', 'B', "C", "D", "E", "F"), nrow = 2,
          align = "h")
```
